requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
chardet==3.0.4
Click==7.0
discord.py==1.2.5
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
google==2.0.3
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.7.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dotenv==0.10.3
soupsieve==1.9.5
SQLAlchemy==1.3.12
websockets==6.0
Werkzeug==0.16.0
yarl==1.4.2
